+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Mob Farming Supplies
 
+## 26.1.2.2
+
+* Add configuration UI to Ender Inhibitor
+* Fix crash with tintSource on Fabric (thanks @jajajafeti)
+
 ## 26.1.2.1
 
 * Add fun buttons to spice up the mob farm